--- a/crates/bevy_pbr/src/light/ambient_light.rs
+++ b/crates/bevy_pbr/src/light/ambient_light.rs
@@ -1,3 +1,8 @@
+#![deprecated(
+    since = "0.16.0",
+    note = "Use `EnvironmentMapLight::solid_color` instead"
+)]
+
 use super::*;
 
 /// An ambient light, which lights the entire scene equally.