--- a/crates/bevy_pbr/src/light/mod.rs
+++ b/crates/bevy_pbr/src/light/mod.rs
@@ -20,7 +20,7 @@ use bevy_render::{
 use bevy_transform::components::{GlobalTransform, Transform};
 use bevy_utils::Parallel;
 
-use crate::*;
+use crate::{prelude::EnvironmentMapLight, *};
 
 mod ambient_light;
 pub use ambient_light::AmbientLight;
@@ -509,6 +509,7 @@ pub struct LightVisibilityClass;
 /// System sets used to run light-related systems.
 #[derive(Debug, Hash, PartialEq, Eq, Clone, SystemSet)]
 pub enum SimulationLightSystems {
+    MapAmbientLights,
     AddClusters,
     AssignLightsToClusters,
     /// System order ambiguities between systems in this set are ignored:
@@ -520,6 +521,33 @@ pub enum SimulationLightSystems {
     /// the order of systems within this set is irrelevant, as the various visibility-checking systems
     /// assumes that their operations are irreversible during the frame.
     CheckLightVisibility,
+}
+
+pub fn map_ambient_lights(
+    mut commands: Commands,
+    mut image_assets: ResMut<Assets<Image>>,
+    ambient_light: Res<AmbientLight>,
+    views: Query<
+        (
+            Entity,
+            Option<Ref<AmbientLight>>,
+            Option<&EnvironmentMapLight>,
+        ),
+        With<Camera>,
+    >,
+) {
+    let ambient_light = ambient_light.into();
+    for (entity, ambient_override, environment_map) in views.iter() {
+        let ambient = ambient_override.as_ref().unwrap_or(&ambient_light);
+        let ambient_required = ambient.brightness > 0.0 && ambient.color != Color::BLACK;
+        if ambient_required && environment_map.is_none() && ambient.is_changed() {
+            commands.entity(entity).insert(EnvironmentMapLight {
+                intensity: ambient.brightness,
+                affects_lightmapped_mesh_diffuse: ambient.affects_lightmapped_meshes,
+                ..EnvironmentMapLight::solid_color(image_assets.as_mut(), ambient.color)
+            });
+        }
+    }
 }
 
 pub fn update_directional_light_frusta(