--- a/crates/bevy_pbr/src/light/mod.rs
+++ b/crates/bevy_pbr/src/light/mod.rs
@@ -527,20 +527,13 @@ pub fn map_ambient_lights(
     mut commands: Commands,
     mut image_assets: ResMut<Assets<Image>>,
     ambient_light: Res<AmbientLight>,
-    views: Query<
-        (
-            Entity,
-            Option<Ref<AmbientLight>>,
-            Option<&EnvironmentMapLight>,
-        ),
-        With<Camera>,
-    >,
+    views: Query<(Entity, Option<Ref<AmbientLight>>), With<Camera>>,
 ) {
     let ambient_light = ambient_light.into();
-    for (entity, ambient_override, environment_map) in views.iter() {
+    for (entity, ambient_override) in views.iter() {
         let ambient = ambient_override.as_ref().unwrap_or(&ambient_light);
         let ambient_required = ambient.brightness > 0.0 && ambient.color != Color::BLACK;
-        if ambient_required && environment_map.is_none() && ambient.is_changed() {
+        if ambient_required && ambient.is_changed() {
             commands.entity(entity).insert(EnvironmentMapLight {
                 intensity: ambient.brightness,
                 affects_lightmapped_mesh_diffuse: ambient.affects_lightmapped_meshes,