--- a/crates/bevy_pbr/src/light/mod.rs
+++ b/crates/bevy_pbr/src/light/mod.rs
@@ -523,22 +523,50 @@ pub enum SimulationLightSystems {
     CheckLightVisibility,
 }
 
+#[derive(Component)]
+pub struct EnvironmentMapLightFromAmbientLight;
+
 pub fn map_ambient_lights(
     mut commands: Commands,
     mut image_assets: ResMut<Assets<Image>>,
     ambient_light: Res<AmbientLight>,
-    views: Query<(Entity, Option<Ref<AmbientLight>>), With<Camera>>,
+    new_views: Query<
+        (Entity, Option<Ref<AmbientLight>>),
+        (
+            With<Camera>,
+            Without<EnvironmentMapLight>,
+            Without<EnvironmentMapLightFromAmbientLight>,
+        ),
+    >,
+    mut managed_views: Query<
+        (&mut EnvironmentMapLight, Option<Ref<AmbientLight>>),
+        With<EnvironmentMapLightFromAmbientLight>,
+    >,
 ) {
     let ambient_light = ambient_light.into();
-    for (entity, ambient_override) in views.iter() {
+    for (entity, ambient_override) in new_views.iter() {
         let ambient = ambient_override.as_ref().unwrap_or(&ambient_light);
         let ambient_required = ambient.brightness > 0.0 && ambient.color != Color::BLACK;
         if ambient_required && ambient.is_changed() {
-            commands.entity(entity).insert(EnvironmentMapLight {
+            commands
+                .entity(entity)
+                .insert(EnvironmentMapLight {
+                    intensity: ambient.brightness,
+                    affects_lightmapped_mesh_diffuse: ambient.affects_lightmapped_meshes,
+                    ..EnvironmentMapLight::solid_color(image_assets.as_mut(), ambient.color)
+                })
+                .insert(EnvironmentMapLightFromAmbientLight);
+        }
+    }
+    for (mut env_map, ambient_override) in managed_views.iter_mut() {
+        let ambient = ambient_override.as_ref().unwrap_or(&ambient_light);
+        let ambient_required = ambient.brightness > 0.0 && ambient.color != Color::BLACK;
+        if ambient_required && ambient.is_changed() {
+            *env_map = EnvironmentMapLight {
                 intensity: ambient.brightness,
                 affects_lightmapped_mesh_diffuse: ambient.affects_lightmapped_meshes,
                 ..EnvironmentMapLight::solid_color(image_assets.as_mut(), ambient.color)
-            });
+            };
         }
     }
 }