--- a/crates/bevy_pbr/src/light/mod.rs
+++ b/crates/bevy_pbr/src/light/mod.rs
@@ -23,7 +23,7 @@ use bevy_utils::Parallel;
 use crate::{prelude::EnvironmentMapLight, *};
 
 mod ambient_light;
-#[allow(deprecated)]
+#[expect(deprecated)]
 pub use ambient_light::AmbientLight;
 
 mod point_light;
@@ -527,7 +527,7 @@ pub enum SimulationLightSystems {
 #[derive(Component)]
 pub struct EnvironmentMapLightFromAmbientLight;
 
-#[allow(deprecated)]
+#[expect(deprecated)]
 pub fn map_ambient_lights(
     mut commands: Commands,
     mut image_assets: ResMut<Assets<Image>>,