--- a/examples/3d/fog.rs
+++ b/examples/3d/fog.rs
@@ -21,6 +21,10 @@ use bevy::{
 };
 
 fn main() {
+    #[expect(
+        deprecated,
+        reason = "Once AmbientLight is removed, the resource can be removed"
+    )]
     App::new()
         .insert_resource(AmbientLight::NONE)
         .add_plugins(DefaultPlugins)