… crates/bevy_pbr/src/render/pbr_ambient.wgsl
… examples/3d/mixed_lighting.rs
… examples/3d/motion_blur.rs
… examples/3d/skybox.rs
… examples/3d/specular_tint.rs
… examples/3d/spherical_area_lights.rs
… examples/3d/spotlight.rs
… examples/3d/ssao.rs
… examples/3d/volumetric_fog.rs
… examples/animation/animated_mesh.rs
… examples/animation/animated_mesh_control.rs
… examples/animation/animated_mesh_events.rs
… examples/animation/animated_transform.rs
… examples/animation/animation_graph.rs
… examples/animation/animation_masks.rs
… examples/animation/custom_skinned_mesh.rs
… examples/animation/gltf_skinned_mesh.rs
… examples/animation/morph_targets.rs
… examples/asset/multi_asset_sync.rs
… examples/math/render_primitives.rs