@SparkyPotato 2025-01-21
> > I can implement hemispherical lights in this PR too if needed
> 
> The PR is simple enough that you could add this in too I think. Up to you, but I would approve it either way.

Might as well add it in then.