@SparkyPotato 2025-01-22
> > The lighting example does go from about 0.51 ms to 0.55 ms on my machine (and enabling deferred makes that 0.46 ms to 0.49 ms)
> 
> is that in release? if so then i'd argue a ~10% fps reduction is not justified, but it's surprising how large the impact is.

In release, the difference is 0.49 ms 0.53 ms, but the deferred timings don't change.

> > They should be considered basically mutually exclusive (strictly speaking, ambient should be considered a type of light probe, which this PR moves us toward).
> 
> makes sense. if the impact is really this big then perhaps we need some investigation into env map perf before going further though?

I'd think ambient lights are usually used only in simple scenes where we'd be CPU bound anyways. For anything more complex, users would probably switch to an environment map or one baked by the procedural sky.