@SparkyPotato 2025-01-23
I decided to profile the case without an ambient light or envmap, which takes 0.47 ms on this branch.

This means we'll actually be slightly _faster_ in scenes with an envmap after this PR. For example: 
- the deferred rendering example goes from 1.16 to 1.12 ms
- SSR goes from 0.92 to 0.89 ms
- post processing goes from 1.02 to 0.96 ms
- meshlet goes from 1.68 to 1.67 ms
- color grading goes from 0.88 to 0.87 ms

Note that these are all overall frametimes and not individual pass timings like that of the lighting example (for lighting, the overall frame went from 0.98 to 0.96 ms without any ambient or env light).