@SparkyPotato 2025-02-06
I think I've addressed all requested changes and updated all the examples to use `EnvironmentMapLight` except when disabling the default `AmbientLight` (which will, of course, no longer be required when we remove `AmbientLight`).