@IceSentry 2025-02-12
I started an example runner run here https://github.com/TheBevyFlock/bevy-example-runner/issues/117

Other than that LGTM. If someone is concerned with the texture sampling for performance reasons they probably want to use a custom material that is just simpler in general too, but I think this setup makes sense for most users of StandardMaterial.